Starting to convert to Bootstrap 5
This is a start at getting the wysiwyg editor working in bootstrap 5. Still need some tweaking for the video, image, tables and probably more. 

Bootstrap 5 is not allowing multiple trigger events on the same element, so having the tooltip on the button with the editors action listeners where preventing most of the script to not work.

Also had to use d-none to hide the editor box instead of invisible.